+++ frontend/eslint.config.mjs
@@ -11,6 +11,17 @@ const compat = new FlatCompat({
 
 const eslintConfig = [
   ...compat.extends("next/core-web-vitals", "next/typescript"),
+  {
+    files: ["**/*.ts", "**/*.tsx"],
+    rules: {
+      "@typescript-eslint/no-unused-vars": "off", // 👈 turns off the error
+      "@typescript-eslint/no-explicit-any": "warn", // maybe just warn instead of error?
+      "react-hooks/rules-of-hooks": "warn", // keep this one strict!
+      "@typescript-eslint/no-wrapper-object-types" : "warn",
+      "@typescript-eslint/no-explicit-any" : "warn"
+
+    },
+  }
 ];
 
 export default eslintConfig;
