+++ frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,61 +1,5 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-
-export const authOptions : NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "text", placeholder: "user@example.com" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials, req) {
-        const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/auth/login`, {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            email: credentials?.email,
-            password: credentials?.password,
-          }),
-        });
-
-        const data = await res.json();
-
-        if (res.ok && data.user && data.token) {
-          // Return both user and token to JWT
-          return { ...data.user, token: data.token };
-        }
-
-        // Authentication failed
-        return null;
-      },
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  pages : {
-    signIn : "/login"
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.user = {
-          _id: user._id,
-          name: user.name!,
-          email: user.email!,
-        };
-        token.accessToken = user.token;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      session.user = token.user!;
-      session.accessToken = token.accessToken;
-      return session;
-    },
-  },
-}
+import { authOptions } from "@/app/lib/auth";
+import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
