+++ frontend/src/app/incomes/page.tsx
@@ -1,41 +1,31 @@
-'use client'
-import React from 'react'
-import PostIncome from './PostIncome'
-import { useSession } from 'next-auth/react'
-/**
- * TODO:
- * fetch Incomes ,
- * show total 
- * expand to show details 
- *    sort
- * 
- *
- * make a POST Income FORM
- * validate
- * post income
- * return success or error . 
- * refresh the get data
- */
+import { getServerSession } from "next-auth";
+import PostIncome from "./PostIncome";
+import { Typography } from "@mui/material";
+import { authOptions } from "../lib/auth";
 
-export const getIncomes = async (session : any) => {
-  const res = await fetch(process.env.NEXT_PUBLIC_BACKEND_URL! + '/api/incomes' , {
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${session.data.accessToken}`,
-    },
-  })
-  const income = await res.json()
-  return income
-} 
-const page = () => {
-  const session = useSession()
-  if (session.data?.accessToken) {
-    const data = getIncomes(session)
-    console.log(data)
+export default async function Page() {
+  const session = await getServerSession(authOptions);
+
+  if (!session?.accessToken) {
+    return <div>ログインが必要です</div>;
   }
-  return (
-    <div>fetch incomes</div>
-  )
-}
 
-export default page
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/incomes`, {
+    headers: {
+      Authorization: `Bearer ${session.accessToken}`,
+    },
+    cache: 'no-store',
+  });
+
+  const incomes = await res.json();
+
+  return (
+    <div>
+      <h2>収入一覧</h2>
+      {incomes.map((income: any) => (
+        <Typography key={income._id} >{income.description} - {income.amount} - {income.currency}</Typography>
+      ))}
+      <PostIncome />
+    </div>
+  );
+}
